Safari-609.1.11.1/LayoutTests/http/tests/loading/preload-script-srcset-ignored.html

<!DOCTYPE html>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=1000"></script>
<script src="http://127.0.0.1:8000/resources/checkPreload.js"></script>
<body>
This test makes sure that the Preload scanner doesn't preload the script resource, since it's defined in a srcset attribute.
<br/>
<script>
checkForPreload("resources/preload-test.jpg", false);
</script>
<script srcset="resources/preload-test.jpg">
</body>
